Count değeri int olarak revize edildi.

+++ SignalR.DtoLayer/BasketDto/CreateBasketDto.cs
@@ -9,7 +9,7 @@ namespace SignalR.DtoLayer.BasketDto
     public class CreateBasketDto
     {
         public decimal Price { get; set; }
-        public decimal Count { get; set; }
+        public int Count { get; set; }
         public decimal TotalPrice { get; set; }
         public int ProductID { get; set; }
         public int MenuTableID { get; set; }
